# lib/api-zod/src/generated/types/createMedicineRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediCare Reminder - Medicine Reminder System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMedicineRequestFrequency } from "./createMedicineRequestFrequency";

export interface CreateMedicineRequest {
  patientId: number;
  name: string;
  dosage: string;
  frequency: CreateMedicineRequestFrequency;
  times: string[];
  notes?: string;
  isActive?: boolean;
}