# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediCare Reminder - Medicine Reminder System API
 * OpenAPI spec version: 0.1.0
 */
import type { SensorEvent } from "./sensorEvent";

export interface DashboardSummary {
  totalPatients: number;
  activeMedicines: number;
  todayEvents: number;
  adherenceRate: number;
  recentEvents: SensorEvent[];
}