# lib/api-zod/src/generated/types/medicine.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediCare Reminder - Medicine Reminder System API
 * OpenAPI spec version: 0.1.0
 */
import type { MedicineFrequency } from "./medicineFrequency";

export interface Medicine {
  id: number;
  patientId: number;
  name: string;
  dosage: string;
  frequency: MedicineFrequency;
  times: string[];
  notes?: string;
  isActive: boolean;
  createdAt: Date;
}